move get_context_data to snapshot_util module
